requirements.txt
agno==1.7.1
beautifulsoup4==4.13.4
docling==2.40.0
google-genai==1.24.0
pandoc==2.4
pypandoc==1.15
python-docx==1.2.0
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
typing-extensions==4.14.1
pathlib